+++ Kernel/Devices/ConsoleDevice.h
@@ -21,7 +21,7 @@ public:
     virtual ~ConsoleDevice() override;
 
     // ^CharacterDevice
-    virtual bool can_read(const Kernel::OpenFileDescription&, size_t) const override;
+    virtual bool can_read(const Kernel::OpenFileDescription&, u64) const override;
     virtual bool can_write(const Kernel::OpenFileDescription&, size_t) const override { return true; }
     virtual ErrorOr<size_t> read(OpenFileDescription&, u64, Kernel::UserOrKernelBuffer&, size_t) override;
     virtual ErrorOr<size_t> write(OpenFileDescription&, u64, const Kernel::UserOrKernelBuffer&, size_t) override;
